Moving manual cases from test_manual and remove those that are automated
